+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 * (bug) Add missing bundle config.
+* (improvement) Add direct getter for `VideoDetails::getEmbedUrl()`.
 
 
 1.0.0